Cleanup entrypoint file with fh-cluster & functions
The previous attempt at implementing a backoff strategy added more
complexity to an already complex startup process.

This new way using fh-cluster is a deal simpler: all of the clustering
and backoff logic is done behind the scenes in fh-cluster, and we just
give it our app function.

This change also restructures a little, by using more named functions,
in the hope that this will make it easier to understand & maintain.

JIRA: https://issues.jboss.org/browse/RHMAP-2459