Adding response status code 500 for health checks of status 'crit'
Motivation:
Currently all health check responses return a HTTP status code of 200.
This can cause issue with clients that expect to receive a non 200 code.
If fh-health returns a healt check with a response status of 'crit' we
should return a HTTP 500 response code.

Modifications:
Added a check for the health check status of 'crit' and set the response
status to 500.

Result:
All test pass:
$ grunt fh:unit

JIRA:
https://issues.jboss.org/browse/RHMAP-4838