Adding FH_LOG_LEVEL to ose-placeholders.js
Motivation:
When deployed to an OpenShift V3 environment we want to be able to
update the configuration properties without having to edit the
configuration file. Adding an environment variable will allow for us to
update the log level.

Modifications:
Added FH_LOG_LEVEL to config/ose-placeholders.js and also
updated the current log level in config/dev.json to be info instead of
trace.

Result:
Running the following command should genereate a conf-docker.json with
the correct placeholder for the logger stream's log levels:
$ grunt fh-generate-dockerised-config
$ grep -e '\<level' conf-docker.json
        "level": "$(FH_LOG_LEVEL)",
        "level": "$(FH_LOG_LEVEL)",

JIRA:
https://issues.jboss.org/browse/RHMAP-5569