Adding PDF export endpoint route to app forms
Motivation:
Enable PDF representations of a submission to be exported via the Cloud
App API.

Modifications:
Added a new route to lib/handlers/app/forms.js. The handler for this
route differs from the existing in that it uses a separate handlers.js
instead of directly passing the request through to fh-forms.

Result:
All unit test pass and there is now an endpoint available to clients
using the following path:
/:domain/:environment/:projectid/:appid/appforms/submission/:id/exportpdf

JIRA:
https://issues.jboss.org/browse/RHMAP-4718