Adding rule to allow unused 'next' argument to functions.
Motivation:
The main motivation for this rule is to allow Express middleware error
handling function to have an additional 'next' argument. This argument
allows for multiple error handlers to be set up and they can invoke the
next handler in the chain. But for handlers that don't have any more
handlers to call ESLint will complain that the 'next' argument is
unused.

Modifications:
Added a rule to ESlint configuration to ignore arguments that are named
'next'.
Added the removed 'next' parameter to the error handler functions.

Result:
grunt eslint passes without any errors/warnings.