Adding eslint to default grunt target.
Motivation:
When adding the ESlint support I incorrectly assumed that grunt-fh-build
would run this as part of the default target.

Modifications:
Added eslint to the default target so that when grunt is run the eslint
checks will be performed.
Also fixed to minor eslit errors that were added since eslint was added.

Result:
ESlint will now be run as part of the default Grunt target.