Adding pre-export checks for submission export
Motivation:
Before exporting a submission there are a number of checks that should
be performed before the export. Some of these steps are shared by the
app data export and the submission export so we should extract the
common functions into a separate model and have both exports use them.

Modifications:
Extracted the functions common to app data export and submission export.
Added a test for the existing preperationSteps.js.

Result:
All tests pass.

JIRA:
https://issues.jboss.org/browse/RHMAP-6637