Fixing minor eslint error in lib/jobs/appDataRunnerJob
Motivation:
This issue should have been picked up by the build server as we added an
explicit task to run eslint as part of the build. When I checked this
setting this morning it was not present. Either I made a mistake when
adding this setting (not saving properly or something) or it was
removed. Either way I've added it now and double checked that it was
saved.

Modifications:
Fixed the issue.

Result:
Running the default grunt task no longer failes with an error.